Refactor: Reorganize project structure
I've decomposed the monolithic `structure_predict.py` file into a more modular and maintainable structure.

The key changes include:
- `data_models.py`: Houses all Pydantic models and Enums used for data representation and LLM interaction.
- `task_management.py`: Contains classes `Task` and `Task_Queue` for defining and managing individual tasks.
- `goal_management.py`: Includes `Goal_Info`, `Snapshot_Node`, and `Research_Snapshot` for managing the hierarchical structure of research goals.
- `research_agent.py`: The original `structure_predict.py` was renamed to this and now primarily contains the `Research_Agent` class, which orchestrates the research process. It imports components from the new modules.

This reorganization improves code clarity, separation of concerns, and overall maintainability without altering the core functionality.
